Ensuring all functions in rolesActions.ts are authenticated

+++ app/(authenticated)/admin/roles/[roleID]/rolesActions.ts
@@ -8,11 +8,13 @@ import { zodErrorResponse } from "@/components/FormServerHelpers";
 import { RoleSchema } from "@/app/(authenticated)/admin/roles/schema";
 import { ExposedUser } from "@/features/people";
 import { UserType } from "@/features/role";
+import { requirePermission } from "@/lib/auth/server";
 
 export async function editRole(
   roleID: number,
   form: z.infer<typeof RoleSchema>,
 ): Promise<FormResponse> {
+  await requirePermission("ManageMembers.Groups");
   const payload = RoleSchema.safeParse(form);
   if (!payload.success) {
     return zodErrorResponse(payload.error);
@@ -24,24 +26,28 @@ export async function editRole(
 }
 
 export async function deleteRole(roleID: number): Promise<FormResponse> {
+  await requirePermission("ManageMembers.Groups");
   await Role.deleteRole(roleID);
   revalidatePath(`/admin/roles`);
   return { ok: true } as const;
 }
 
 export async function addUserToRole(roleID: number, userID: number) {
+  await requirePermission("ManageMembers.Groups");
   await Role.addUserToRole(roleID, userID);
   revalidatePath(`/admin/roles/${roleID}`);
   return { ok: true } as const;
 }
 
 export async function removeUserFromRole(roleID: number, userID: number) {
+  await requirePermission("ManageMembers.Groups");
   await Role.removeUserFromRole(roleID, userID);
   revalidatePath(`/admin/roles/${roleID}`);
   return { ok: true } as const;
 }
 
 export async function addPermissionToRole(roleID: number, permission: string) {
+  await requirePermission("ManageMembers.Groups");
   await Role.addPermissionToRole(roleID, permission);
   revalidatePath(`/admin/roles/${roleID}`);
   return { ok: true } as const;
@@ -51,6 +57,7 @@ export async function removePermissionFromRole(
   roleID: number,
   permission: string,
 ) {
+  await requirePermission("ManageMembers.Groups");
   await Role.removePermissionFromRole(roleID, permission);
   revalidatePath(`/admin/roles/${roleID}`);
   return { ok: true } as const;
